refactor: extracting device, creating notification for device support warning
