+++ src/style/style.css
@@ -454,7 +454,7 @@ a {
 }
 
 .ant-card-hoverable:hover {
-    border-color: #4b69ab !important;
+    border-color: #45729f !important;
 }
 
 .custom-dropdown-button .ant-btn-compact-first-item {
