+++ src/types/Form.d.ts
@@ -50,8 +50,8 @@ export type Form = {
   PRIMARY_ELEMENTAL_DMG_TYPE: string
   statSim?: any
   resultSort?: string
-  path?: string
   resultsLimit?: number
+  path?: string
 
   weights: {
     [key: string]: number
